Re #48 manually handle signals in wobbler example
This commit removes reliance on rospy for handling the shutdown
behavior when a Ctrl+C signal is asserted on the joint_velocity_wobbler
example. Cleanup was not properly being executed in only the
simulated robot through rospy's on_shutdown() callback handler. This
change has been tested on both the simulated and real Baxter robots.